+++ example/index.js
@@ -3,7 +3,7 @@ const {
   JsonBody,
   Routex,
   TextBody,
-  useExpress
+  useExpress,
 } = require("../dist/lib");
 
 class ValidationError extends ErrorWithStatusCode {
@@ -18,16 +18,16 @@ const port = process.env.PORT || 3000;
 const app = new Routex();
 
 app
-  .get("/", async ctx => {
+  .get("/", async (ctx) => {
     ctx.body = new TextBody("Index");
   })
   .get("/error", [
     () => {
       throw new ValidationError("Some error");
     },
-    ctx => {
+    (ctx) => {
       ctx.body = new TextBody("Will never be returned");
-    }
+    },
   ])
   .get(
     "/express",
@@ -35,13 +35,13 @@ app
       res.write("Express!");
     })
   )
-  .get("/delay", async ctx => {
-    await new Promise(resolve => setTimeout(resolve, 500));
+  .get("/delay", async (ctx) => {
+    await new Promise((resolve) => setTimeout(resolve, 500));
     ctx.body = new TextBody("Delayed");
   })
   .get(
     "/catch",
-    ctx => {
+    (ctx) => {
       ctx.body = new TextBody("Catch");
     },
     { exact: false }
@@ -49,14 +49,14 @@ app
 
 app
   .child("/child")
-  .middleware(ctx => {
+  .middleware((ctx) => {
     ctx.res.setHeader("X-Server", "Routex");
     ctx.data.name = "john";
   })
-  .get("/", ctx => {
+  .get("/", (ctx) => {
     ctx.body = new TextBody("Child index");
   })
-  .get("/json", ctx => {
+  .get("/json", (ctx) => {
     ctx.body = new JsonBody({ child: true, name: ctx.data.name });
   });