+++ src/express.ts
@@ -2,9 +2,9 @@
 import { Handler } from "./handler";
 
 export function useExpress(handler: any): Handler {
-  return async ctx => {
+  return async (ctx) => {
     if (handler.length >= 3) {
-      return new Promise<void>(resolve => {
+      return new Promise<void>((resolve) => {
         handler(ctx.req, ctx.res, resolve);
       });
     } else {